refactor: move tests directly into `pterm` module
