+++ internal/capture_stdout.go
@@ -1,4 +1,4 @@
-package test
+package internal
 
 import (
 	"io"
@@ -8,8 +8,8 @@ import (
 	"github.com/gookit/color"
 )
 
-// captureStdout captures everything written to the terminal and returns it as a string.
-func captureStdout(f func(w io.Writer)) string {
+// CaptureStdout captures everything written to the terminal and returns it as a string.
+func CaptureStdout(f func(w io.Writer)) string {
 	originalStdout := os.Stdout
 	r, w, _ := os.Pipe()
 	os.Stdout = w
