--- test/collection.go
@@ -1,4 +1,0 @@
-package test
-
-var randomStrings = []string{"hello world", "²³14234!`§=)$-.€@_&", "This is a sentence.",
-	"This\nstring\nhas\nmultiple\nlines", "windows\r\nline\r\nendings", "\rtext"}
